+++ src/refactoring/entities/product/ui/index.ts
@@ -1,2 +1,1 @@
 export { default as ProductInfoCard } from "./ProductInfoCard";
-export { default as CreateProductForm } from "./CreateProductForm";
