+++ src/refactoring/features/product/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-product";
