+++ src/refactoring/features/product/model/use-product.ts
