+++ src/refactoring/features/product/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateProductForm } from "./CreateProductForm";
